pyproject.toml
[project]
name = "cherry-ai-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cherry-etl>=0.0.5",
    "langchain[openai]>=0.3.20",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff]
line-length = 100
indent-width = 4
